Now the user is correctly notified when the imported books does not match the canon specification, and also, the canon specification automatically matches the file imported
