+++ renderer/src/layouts/editor/ConfirmationModal.js
@@ -16,6 +16,7 @@ export default function ConfirmationModal({
     active: false,
     loading: false,
   },
+  showCancelButton=true,
 }) {
   const cancelButtonRef = useRef(null);
   const { t } = useTranslation();
@@ -97,14 +98,14 @@ export default function ConfirmationModal({
                   {buttonName}
                 </button>
               ) }
-                <button
+                {showCancelButton && <button
                   type="button"
                   className="mt-3 w-full inline-flex justify-center rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-base font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
                   onClick={() => setOpenModal(false)}
                   ref={cancelButtonRef}
                 >
                   {t('btn-cancel')}
-                </button>
+                </button>}
               </div>
             </div>
           </Transition.Child>
